feat(dashboard): wire Source Activity panel for Remote MCP sources https://linear.app/speakeasy/issue/AGE-2214/source-activity-panel-for-remote-mcp-server-sources
